Split profile
