+++ src/main/java/eu/dissco/core/datacitepublisher/Profiles.java
@@ -9,4 +9,5 @@ public class Profiles {
   }
 
   public static final String TEST = "test";
+  public static final String PUBLISH = "publish";
 }
