+++ src/main/java/eu/dissco/core/datacitepublisher/controller/RecoveryController.java
@@ -1,12 +1,14 @@
 package eu.dissco.core.datacitepublisher.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import eu.dissco.core.datacitepublisher.Profiles;
 import eu.dissco.core.datacitepublisher.domain.RecoveryEvent;
 import eu.dissco.core.datacitepublisher.exceptions.DataCiteApiException;
 import eu.dissco.core.datacitepublisher.exceptions.HandleResolutionException;
 import eu.dissco.core.datacitepublisher.service.RecoveryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/datacite-recovery")
 @RequiredArgsConstructor
 @Slf4j
+@Profile(Profiles.PUBLISH)
 public class RecoveryController {
 
   private final RecoveryService recoveryService;
