+++ src/main/java/eu/dissco/core/datacitepublisher/web/DataCiteClient.java
@@ -3,6 +3,7 @@ package eu.dissco.core.datacitepublisher.web;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import eu.dissco.core.datacitepublisher.Profiles;
 import eu.dissco.core.datacitepublisher.domain.datacite.DcAttributes;
 import eu.dissco.core.datacitepublisher.exceptions.DataCiteApiException;
 import java.time.Duration;
@@ -10,6 +11,7 @@ import java.util.concurrent.ExecutionException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.BodyInserters;
@@ -20,9 +22,10 @@ import reactor.util.retry.Retry;
 @RequiredArgsConstructor
 @Component
 @Slf4j
+@Profile(Profiles.PUBLISH)
 public class DataCiteClient {
 
-  @Qualifier(value = "datacite")
+  @Qualifier("datacite")
   private final WebClient webClient;
   @Qualifier("objectMapper")
   private final ObjectMapper mapper;
